Split 'utils.py' in different modules inside the new package 'lib'. Remove 'lib' from gitignore. Add more info when the download data script fails
