+++ albam/engines/mtframework/utils.py
@@ -6,9 +6,7 @@ from albam.exceptions import BuildMeshError
 from albam.engines.mtframework.mod import (
     VERTEX_FORMATS_TO_CLASSES,
     )
-from albam.utils import (
-    get_size,
-    )
+from albam.lib.structure import get_size
 
 
 def get_vertices_array(mod, mesh):
